pyproject.toml
[tool.poetry]
name = "qa-guru-diploma-swaggerpetstore-api"
version = "0.1.0"
description = ""
authors = ["Irina.Voron <yarutich.irina@gmail.com>"]
readme = "README.md"
packages = [{include = "qa"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
jsonschema = "^4.22.0"
pytest = "^8.2.0"
selene = "2.0.0rc9"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
allure-pytest = "^2.13.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
